修复一些bug

+++ 简易示波器-最终案例/Oscilloscope/User/main.h
@@ -10,7 +10,7 @@ struct Oscilloscope
    uint16_t pwmOut;         //PWM引脚输出的PWM占空比
    uint32_t sampletime;     //电压采集时间
    uint32_t timerPeriod;    //定时器周期设置
-   float gatherFreq;        //示波器采集频率
+   uint32_t gatherFreq;        //示波器采集频率
    float vpp;               //峰峰值
    float voltageValue[300]; //ADC采集电压值
 };
